Refactor allpass classes (favor composition over inheritance)

+++ src/Buffer.cpp
@@ -1,4 +1,8 @@
 #include "Buffer.hpp"
+#include "utility.hpp"
+#include <cassert>
+#include <cstring>
+#include <new>
 
 using md_audio::Buffer;
 using md_audio::MdFloat;
@@ -17,13 +21,13 @@ void Buffer::initialise() {
     std::memset(m_memory, 0, m_size * sizeof(MdFloat));
 }
 
-MdFloat& Buffer::operator[] (std::size_t index) noexcept {
+MdFloat& Buffer::operator[] (std::uint32_t index) noexcept {
     assert(index >= 0 && index < m_size);
 
     return m_memory[index];
 }
 
-const MdFloat& Buffer::operator[] (std::size_t index) const noexcept {
+const MdFloat& Buffer::operator[] (std::uint32_t index) const noexcept {
     assert(index >= 0 && index < m_size);
 
     return m_memory[index];
